+++ robot_descriptions/_package_dirs.py
@@ -7,6 +7,8 @@
 import os.path
 from typing import List
 
+from ._xacro import get_urdf_path
+
 
 def get_package_dirs(module) -> List[str]:
     """Get package directories for a given module.
@@ -17,10 +19,14 @@ def get_package_dirs(module) -> List[str]:
     Returns:
         Package directories.
     """
-    return [
+    package_dirs = [
         module.PACKAGE_PATH,
         module.REPOSITORY_PATH,
         os.path.dirname(module.PACKAGE_PATH),
         os.path.dirname(module.REPOSITORY_PATH),
-        os.path.dirname(module.URDF_PATH),  # e.g. laikago_description
     ]
+    if hasattr(module, "URDF_PATH") or hasattr(module, "XACRO_PATH"):
+        package_dirs.append(
+            os.path.dirname(get_urdf_path(module))  # e.g. laikago_description
+        )
+    return package_dirs